requirements.txt
Flask
numpy
scikit-learn
matplotlib
Pillow
